fixing some warnings
git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@3666 a27c51de-2d2b-4d1b-a6e8-f7f87814f453

+++ projects/BinQ/BinQSupport.C
@@ -108,7 +108,7 @@ std::string BinQSupport::ToUpper(std::string myString)
 
 SgNode* BinQSupport::disassembleFile(std::string tsv_directory, std::string& sourceFile){
   SgNode* globalBlock=NULL;
-  int found = tsv_directory.rfind(".");
+  size_t found = tsv_directory.rfind(".");
   string ending="";
 
   if (found!=string::npos) {
